Update multi modal embeddings with NomicEmbedding
