--- llama-index-integrations/embeddings/llama-index-embeddings-nomic-multimodal/BUILD
@@ -1,3 +1,0 @@
-poetry_requirements(
-    name="poetry",
-)
