--- llama-index-integrations/embeddings/llama-index-embeddings-nomic-multimodal/README.md
@@ -1,1 +1,0 @@
-# LlamaIndex Embeddings Integration: Nomic MultiModal
