--- llama-index-integrations/embeddings/llama-index-embeddings-nomic-multimodal/llama_index/embeddings/nomic_multimodal/BUILD
@@ -1,1 +1,0 @@
-python_sources()
