--- llama-index-integrations/embeddings/llama-index-embeddings-nomic-multimodal/llama_index/embeddings/nomic_multimodal/__init__.py
@@ -1,3 +1,0 @@
-from llama_index.embeddings.nomic_multimodal.base import NomicMultiModalEmbedding
-
-__all__ = ["NomicMultiModalEmbedding"]
