--- llama-index-integrations/embeddings/llama-index-embeddings-nomic-multimodal/tests/BUILD
@@ -1,1 +1,0 @@
-python_tests()
